startup setup prepared
