+++ templates/item_header.html
@@ -9,5 +9,5 @@
         <td>Owned:</td>
         <td>Played:</td>
         <td>Play dates:</td>
-        <td>Comments</td>
+        <td>Comments:</td>
 </tr>
